fix random_int function returning int instead of long; add type_of function

+++ src/main/java/net/darmo_creations/mccode/interpreter/builtin_functions/RandomIntFunction.java
@@ -36,6 +36,6 @@ public class RandomIntFunction extends BuiltinFunction {
     if (a > b) {
       throw new EvaluationException(scope, "mccode.interpreter.error.random_invalid_bounds", a, b);
     }
-    return scope.getProgram().getRNG().nextInt(b - a) + a;
+    return (long) (scope.getProgram().getRNG().nextInt(b - a) + a);
   }
 }
